% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnegie-package.R
\docType{data}
\name{region}
\alias{region}
\title{A Lookup Table of State Abbreviations and Regions}
\format{A data frame with 60 rows and 3 variables}
\usage{
data(region)
}
\description{
A dataset containing Carnegie region classifications and corresponding states.
}
\details{
\itemize{
  \item ID. An identifier for the region + state
  \item Region. A region of states
  \item State. A state abbreviation
}
}
\references{
\url{http://carnegieclassifications.iu.edu}
}
\keyword{datasets}
